file renaming and small updates
